Split EAV code into multiple files
