+++ core_types/src/eav/eavi.rs
@@ -10,16 +10,13 @@ use crate::{
     json::JsonString,
 };
 use chrono::offset::Utc;
-use objekt;
-use std::{
-    cmp::Ordering,
-    collections::BTreeSet,
-    convert::TryInto,
-    sync::{Arc, RwLock},
+use eav::{
+    query::{EaviQuery, IndexFilter},
+    storage::{EntityAttributeValueStorage, ExampleEntityAttributeValueStorage},
 };
+use std::{cmp::Ordering, collections::BTreeSet, convert::TryInto};
 
 use regex::RegexBuilder;
-use std::fmt::Debug;
 
 /// Address of AddressableContent representing the EAV entity
 pub type Entity = Address;
@@ -59,156 +56,6 @@ impl Ord for EntityAttributeValueIndex {
     fn cmp(&self, other: &EntityAttributeValueIndex) -> Ordering {
         self.index.cmp(&other.index())
     }
-}
-
-/// Represents a set of filtering operations on the EAVI store.
-pub struct EaviQuery<'a> {
-    entity: EntityFilter<'a>,
-    attribute: AttributeFilter<'a>,
-    value: ValueFilter<'a>,
-    index: IndexFilter,
-}
-
-type EntityFilter<'a> = EavFilter<'a, Entity>;
-type AttributeFilter<'a> = EavFilter<'a, Attribute>;
-type ValueFilter<'a> = EavFilter<'a, Value>;
-
-impl<'a> Default for EaviQuery<'a> {
-    fn default() -> EaviQuery<'a> {
-        EaviQuery::new(
-            Default::default(),
-            Default::default(),
-            Default::default(),
-            IndexFilter::LatestByAttribute,
-        )
-    }
-}
-
-impl<'a> EaviQuery<'a> {
-    pub fn new(
-        entity: EntityFilter<'a>,
-        attribute: AttributeFilter<'a>,
-        value: ValueFilter<'a>,
-        index: IndexFilter,
-    ) -> Self {
-        Self {
-            entity,
-            attribute,
-            value,
-            index,
-        }
-    }
-
-    pub fn run<I>(&self, iter: I) -> BTreeSet<EntityAttributeValueIndex>
-    where
-        I: Clone + Iterator<Item = EntityAttributeValueIndex> + 'a,
-    {
-        let iter2 = iter.clone();
-        let filtered = iter
-            .clone()
-            .filter(|eavi| EaviQuery::eav_check(&eavi, &self.entity, &self.attribute, &self.value));
-
-        match self.index {
-            IndexFilter::LatestByAttribute => filtered
-                .filter(|eavi| {
-                    iter2
-                        .clone()
-                        .filter(|eavi_inner| {
-                            EaviQuery::eav_check(
-                                &eavi_inner,
-                                &Some(eavi.entity()).into(),
-                                &self.attribute,
-                                &Some(eavi.value()).into(),
-                            )
-                        })
-                        .last()
-                        .map(|latest| latest.index() == eavi.index())
-                        .unwrap_or(false)
-                })
-                .collect(),
-            IndexFilter::Range(start, end) => filtered
-                .filter(|eavi| {
-                    start.map(|lo| lo <= eavi.index()).unwrap_or(true)
-                        && end.map(|hi| eavi.index() <= hi).unwrap_or(true)
-                })
-                .collect(),
-        }
-    }
-
-    fn eav_check(
-        eavi: &EntityAttributeValueIndex,
-        e: &EntityFilter<'a>,
-        a: &AttributeFilter<'a>,
-        v: &ValueFilter<'a>,
-    ) -> bool {
-        e.check(eavi.entity()) && a.check(eavi.attribute()) && v.check(eavi.value())
-    }
-
-    pub fn entity(&self) -> &EntityFilter<'a> {
-        &self.entity
-    }
-    pub fn attribute(&self) -> &AttributeFilter<'a> {
-        &self.attribute
-    }
-    pub fn value(&self) -> &ValueFilter<'a> {
-        &self.value
-    }
-    pub fn index(&self) -> &IndexFilter {
-        &self.index
-    }
-}
-
-pub struct EavFilter<'a, T: 'a + Eq>(Box<dyn Fn(T) -> bool + 'a>);
-
-impl<'a, T: 'a + Eq> EavFilter<'a, T> {
-    pub fn single(val: T) -> Self {
-        Self(Box::new(move |v| v == val))
-    }
-
-    pub fn attribute_prefixes(
-        prefixes: Vec<&'a str>,
-        base: Option<&'a str>,
-    ) -> AttributeFilter<'a> {
-        Self(Box::new(move |v: Attribute| match base {
-            Some(base) => prefixes
-                .iter()
-                .map(|p| p.to_string() + base)
-                .any(|attr| v.to_owned() == attr),
-            None => prefixes.iter().any(|prefix| v.starts_with(prefix)),
-        }))
-    }
-
-    pub fn predicate<F>(predicate: F) -> Self
-    where
-        F: Fn(T) -> bool + 'a,
-    {
-        Self(Box::new(predicate))
-    }
-
-    pub fn check(&self, val: T) -> bool {
-        self.0(val)
-    }
-}
-
-impl<'a, T: Eq> Default for EavFilter<'a, T> {
-    fn default() -> EavFilter<'a, T> {
-        Self(Box::new(|_| true))
-    }
-}
-
-impl<'a, T: Eq> From<Option<T>> for EavFilter<'a, T> {
-    fn from(val: Option<T>) -> EavFilter<'a, T> {
-        val.map(|v| EavFilter::single(v)).unwrap_or_default()
-    }
-}
-
-/// Specifies options for filtering on Index:
-/// LatestByAttribute is a special kind of lookup used for links. TODO: describe in words
-/// Range returns all results within a particular range of indices.
-#[derive(Clone, Debug)]
-pub enum IndexFilter {
-    LatestByAttribute,
-    Range(Option<i64>, Option<i64>),
 }
 
 impl AddressableContent for EntityAttributeValueIndex {
@@ -282,94 +129,6 @@ impl EntityAttributeValueIndex {
 
     pub fn set_index(&mut self, new_index: i64) {
         self.index = new_index
-    }
-}
-
-/// This provides a simple and flexible interface to define relationships between AddressableContent.
-/// It does NOT provide storage for AddressableContent.
-/// Use cas::storage::ContentAddressableStorage to store AddressableContent.
-pub trait EntityAttributeValueStorage: objekt::Clone + Send + Sync + Debug {
-    /// Adds the given EntityAttributeValue to the EntityAttributeValueStorage
-    /// append only storage.
-    fn add_eavi(
-        &mut self,
-        eav: &EntityAttributeValueIndex,
-    ) -> Result<Option<EntityAttributeValueIndex>, HolochainError>;
-
-    /// Fetch the set of EntityAttributeValues that match constraints according to the latest hash version
-    /// - None = no constraint
-    /// - Some(Entity) = requires the given entity (e.g. all a/v pairs for the entity)
-    /// - Some(Attribute) = requires the given attribute (e.g. all links)
-    /// - Some(Value) = requires the given value (e.g. all entities referencing an Address)
-    fn fetch_eavi(
-        &self,
-        query: &EaviQuery,
-    ) -> Result<BTreeSet<EntityAttributeValueIndex>, HolochainError>;
-
-    // @TODO: would like to do this, but can't because of the generic type param
-    // fn iter<I>(&self) -> I
-    // where
-    //     I: Iterator<Item = EntityAttributeValueIndex>;
-}
-
-clone_trait_object!(EntityAttributeValueStorage);
-
-#[derive(Clone, Debug)]
-pub struct ExampleEntityAttributeValueStorage {
-    storage: Arc<RwLock<BTreeSet<EntityAttributeValueIndex>>>,
-}
-impl ExampleEntityAttributeValueStorage {
-    pub fn new() -> ExampleEntityAttributeValueStorage {
-        ExampleEntityAttributeValueStorage {
-            storage: Arc::new(RwLock::new(BTreeSet::new())),
-        }
-    }
-}
-
-pub fn increment_key_till_no_collision(
-    mut eav: EntityAttributeValueIndex,
-    map: BTreeSet<EntityAttributeValueIndex>,
-) -> HcResult<EntityAttributeValueIndex> {
-    if map
-        .iter()
-        .filter(|e| e.index == eav.index())
-        .collect::<BTreeSet<&EntityAttributeValueIndex>>()
-        .len()
-        > 0
-    {
-        let timestamp = eav.clone().index + 1;
-        eav.set_index(timestamp);
-        increment_key_till_no_collision(eav, map)
-    } else {
-        Ok(eav)
-    }
-}
-impl EntityAttributeValueStorage for ExampleEntityAttributeValueStorage {
-    fn add_eavi(
-        &mut self,
-        eav: &EntityAttributeValueIndex,
-    ) -> Result<Option<EntityAttributeValueIndex>, HolochainError> {
-        let mut map = self.storage.write()?;
-        let new_eav = increment_key_till_no_collision(eav.clone(), map.clone())?;
-        map.insert(new_eav.clone());
-        Ok(Some(new_eav.clone()))
-    }
-
-    fn fetch_eavi(
-        &self,
-        query: &EaviQuery,
-    ) -> Result<BTreeSet<EntityAttributeValueIndex>, HolochainError> {
-        let lock = self.storage.read()?;
-        let set = (*lock).clone();
-        let iter = set.iter().cloned();
-        Ok(query.run(iter))
-    }
-}
-
-impl PartialEq for EntityAttributeValueStorage {
-    fn eq(&self, other: &EntityAttributeValueStorage) -> bool {
-        let query = EaviQuery::default();
-        self.fetch_eavi(&query) == other.fetch_eavi(&query)
     }
 }
