+++ core_types/src/eav/mod.rs
@@ -1,2 +1,5 @@
 pub mod eavi;
-pub use self::eavi::*;
+pub mod query;
+pub mod storage;
+
+pub use self::{eavi::*, query::*, storage::*};
